dom/media/test/test_webvtt_disabled.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=833386
-->
<head>
  <meta charset='utf-8'>
  <title>Test for Bug 833386 - TextTrackList Pref Off</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set": [["media.webvtt.enabled", false]]},
  function() {
    var video = document.createElement("video");
    is(video.textTracks, undefined, "HTMLMediaElement::TextTracks() function should be unavailable.");
    is(video.addTextTrack, undefined, "HTMLMediaElement::AddTextTrack() function should be unavailable.");
    is(typeof TextTrackList, "undefined", "TextTrackList should not be available.");
    is(typeof TextTrack, "undefined", "TextTrack should not be available.");
    is(typeof VTTCue, "undefined", "VTTCue should not be available.");
    is(typeof TextTrackCueList, "undefined", "TextTrackCueList should not be available.");
    SimpleTest.finish();
  }
);
</script>
</pre>
</body>
</html>
